requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.1
Flask-JWT-Extended==4.2.1
python-dateutil==2.8.1
pytz==2021.1
SQLAlchemy==1.4.19
Werkzeug==2.2.3
rich
python-dotenv==0.20.0
marshmallow
Flask-Mail==0.9.1